# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AcousticTaskManifest": {
        "additionalProperties": false,
        "properties": {
          "complete": {
            "const": true,
            "default": true,
            "title": "Complete",
            "type": "boolean"
          },
          "core_version": {
            "title": "Core Version",
            "type": "string"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/AcousticManagedFile"
            },
            "maxItems": 3001,
            "minItems": 1,
            "title": "Files",
            "type": "array"
          },
          "kind": {
            "const": "managed_acoustic_files",
            "default": "managed_acoustic_files",
            "title": "Kind",
            "type": "string"
          },
          "operation": {
            "enum": [
              "acoustic_analysis",
              "textgrid_segment"
            ],
            "title": "Operation",
            "type": "string"
          },
          "policy_version": {
            "default": 1,
            "enum": [
              1,
              2
            ],
            "title": "Policy Version",
            "type": "integer"
          }
        },
        "required": [
          "operation",
          "core_version",
          "files"
        ],
        "title": "AcousticTaskManifest",
        "type": "object"
      },
...
      "AssetList": {
        "properties": {
          "assets": {
            "items": {
              "$ref": "#/components/schemas/AssetView"
            },
            "title": "Assets",
            "type": "array"
          }
        },
        "required": [
          "assets"
        ],
        "title": "AssetList",
        "type": "object"
      },
      "AssetView": {
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "number"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "expected_bytes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Bytes"
          },
          "expires_at": {
            "title": "Expires At",
            "type": "number"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "input",
              "result",
              "archive",
              "temporary"
            ],
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "policy_version": {
            "default": 1,
            "enum": [
              1,
              2
            ],
            "title": "Policy Version",
            "type": "integer"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "reserved_bytes": {
            "title": "Reserved Bytes",
            "type": "integer"
          },
          "sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sha256"
          },
          "size_bytes": {
            "title": "Size Bytes",
            "type": "integer"
          },
          "state": {
            "enum": [
              "uploading",
              "ready",
              "deleting",
              "delete_failed",
              "deleted"
            ],
            "title": "State",
            "type": "string"
          }
        },
        "required": [
          "id",
          "project_id",
          "name",
          "kind",
          "state",
          "size_bytes",
          "reserved_bytes",
          "expected_bytes",
          "sha256",
          "created_at",
          "expires_at",
          "error_code"
        ],
        "title": "AssetView",
        "type": "object"
      },
...
      "EggManifest": {
        "additionalProperties": false,
        "properties": {
          "complete": {
            "const": true,
            "default": true,
            "title": "Complete",
            "type": "boolean"
          },
          "core_version": {
            "title": "Core Version",
            "type": "string"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/AcousticManagedFile"
            },
            "maxItems": 5,
            "minItems": 2,
            "title": "Files",
            "type": "array"
          },
          "kind": {
            "const": "managed_egg_files",
            "default": "managed_egg_files",
            "title": "Kind",
            "type": "string"
          },
          "operation": {
            "const": "egg_analysis",
            "default": "egg_analysis",
            "title": "Operation",
            "type": "string"
          },
          "policy_version": {
            "default": 1,
            "enum": [
              1,
              2
            ],
            "title": "Policy Version",
            "type": "integer"
          }
        },
        "required": [
          "core_version",
          "files"
        ],
        "title": "EggManifest",
        "type": "object"
      },
...
      "FileConfig": {
        "additionalProperties": false,
        "properties": {
          "inputs": {
            "items": {
              "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
              "type": "string"
            },
            "maxItems": 16,
            "title": "Inputs",
            "type": "array"
          },
          "max_output_bytes": {
            "default": 16777216,
            "maximum": 1000000000.0,
            "minimum": 1.0,
            "title": "Max Output Bytes",
            "type": "integer"
          },
          "probe_bytes": {
            "default": 16384,
            "maximum": 33554432.0,
            "minimum": 1.0,
            "title": "Probe Bytes",
            "type": "integer"
          },
          "probe_files": {
            "default": 2,
            "maximum": 4.0,
            "minimum": 1.0,
            "title": "Probe Files",
            "type": "integer"
          }
        },
        "title": "FileConfig",
        "type": "object"
      },
      "FileJobInput": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "$ref": "#/components/schemas/FileConfig"
          },
          "idempotency_key": {
            "pattern": "^[A-Za-z0-9_-]{8,80}$",
            "title": "Idempotency Key",
            "type": "string"
          },
          "operation": {
            "enum": [
              "storage_check",
              "archive_zip",
              "extract_zip"
            ],
            "title": "Operation",
            "type": "string"
          },
          "project_id": {
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "title": "Project Id",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "idempotency_key",
          "operation"
        ],
        "title": "FileJobInput",
        "type": "object"
      },
      "FileManifest": {
        "additionalProperties": false,
        "properties": {
          "complete": {
            "const": true,
            "default": true,
            "title": "Complete",
            "type": "boolean"
          },
          "core_version": {
            "title": "Core Version",
            "type": "string"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/ResultFile"
            },
            "maxItems": 16,
            "minItems": 1,
            "title": "Files",
            "type": "array"
          },
          "kind": {
            "const": "managed_files",
            "default": "managed_files",
            "title": "Kind",
            "type": "string"
          },
          "policy_version": {
            "default": 1,
            "enum": [
              1,
              2
            ],
            "title": "Policy Version",
            "type": "integer"
          }
        },
        "required": [
          "files",
          "core_version"
        ],
        "title": "FileManifest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "JobInput": {
        "additionalProperties": false,
        "properties": {
          "config": {
            "$ref": "#/components/schemas/ProbeConfig"
          },
          "idempotency_key": {
            "pattern": "^[A-Za-z0-9_-]{8,80}$",
            "title": "Idempotency Key",
            "type": "string"
          },
          "operation": {
            "const": "pipeline_check",
            "default": "pipeline_check",
            "title": "Operation",
            "type": "string"
          },
          "project_id": {
            "pattern": "^[0-9a-f]{8}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{4}-[0-9a-f]{12}$",
            "title": "Project Id",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "idempotency_key"
        ],
        "title": "JobInput",
        "type": "object"
      },
      "JobList": {
        "additionalProperties": false,
        "properties": {
          "jobs": {
            "items": {
              "$ref": "#/components/schemas/JobView"
            },
            "title": "Jobs",
            "type": "array"
          }
        },
        "required": [
          "jobs"
        ],
        "title": "JobList",
        "type": "object"
      },
      "JobManifest": {
        "additionalProperties": false,
        "properties": {
          "complete": {
            "const": true,
            "default": true,
            "title": "Complete",
            "type": "boolean"
          },
          "core_version": {
            "title": "Core Version",
            "type": "string"
          },
          "kind": {
            "const": "pipeline_check_metadata",
            "default": "pipeline_check_metadata",
            "title": "Kind",
            "type": "string"
          },
          "sample_count": {
            "maximum": 4096,
            "minimum": 1,
            "title": "Sample Count",
            "type": "integer"
          },
          "sha256": {
            "pattern": "^[0-9a-f]{64}$",
            "title": "Sha256",
            "type": "string"
          }
        },
        "required": [
          "sha256",
          "sample_count",
          "core_version"
        ],
        "title": "JobManifest",
        "type": "object"
      },
      "JobView": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "number"
          },
          "error_code": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Code"
          },
          "generation": {
            "title": "Generation",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "operation": {
            "default": "pipeline_check",
            "enum": [
              "pipeline_check",
              "storage_check",
              "archive_zip",
              "extract_zip",
              "acoustic_analysis",
              "textgrid_segment",
              "spectrogram_to_audio",
              "egg_analysis",
              "lpc_analysis",
              "pitch_manipulation",
              "phonology_induction",
              "speech_synthesis",
              "phonation_synthesis",
              "mfa_alignment",
              "lip_analysis"
            ],
            "title": "Operation",
            "type": "string"
          },
          "progress": {
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Progress",
            "type": "number"
          },
          "project_id": {
            "title": "Project Id",
            "type": "string"
          },
          "result_manifest": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobManifest"
              },
              {
                "$ref": "#/components/schemas/FileManifest"
              },
              {
                "$ref": "#/components/schemas/AcousticTaskManifest"
              },
              {
                "$ref": "#/components/schemas/Spec2WavManifest"
              },
              {
                "$ref": "#/components/schemas/EggManifest"
              },
              {
                "$ref": "#/components/schemas/LpcManifest"
              },
              {
                "$ref": "#/components/schemas/M08Manifest"
              },
              {
                "$ref": "#/components/schemas/M14Manifest"
              },
              {
                "$ref": "#/components/schemas/M06Manifest"
              },
              {
                "$ref": "#/components/schemas/M07Manifest"
              },
              {
                "$ref": "#/components/schemas/M11Manifest"
              },
              {
                "$ref": "#/components/schemas/M05Manifest"
              },
              {
                "type": "null"
              }
            ],
            "title": "Result Manifest"
          },
          "retry_of": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retry Of"
          },
          "state": {
            "enum": [
              "queued",
              "running",
              "cancel_requested",
              "cancelled",
              "failed",
              "interrupted",
              "succeeded"
            ],
            "title": "State",
            "type": "string"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "number"
          },
          "waiting_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Waiting Reason"
          }
        },
        "required": [
          "id",
          "project_id",
          "state",
          "progress",
          "generation",
          "created_at",
          "updated_at",
          "result_manifest",
          "error_code"
        ],
        "title": "JobView",
        "type": "object"
      },
...
      "ProbeConfig": {
        "additionalProperties": false,
        "properties": {
          "sample_count": {
            "default": 4096,
            "maximum": 4096.0,
            "minimum": 1.0,
            "title": "Sample Count",
            "type": "integer"
          },
          "seed": {
            "default": 0,
            "maximum": 255.0,
            "minimum": 0.0,
            "title": "Seed",
            "type": "integer"
          }
        },
        "title": "ProbeConfig",
        "type": "object"
      },
      "ProjectInput": {
        "additionalProperties": false,
        "properties": {
          "name": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "ProjectInput",
        "type": "object"
      },
      "ProjectList": {
        "additionalProperties": false,
        "properties": {
          "projects": {
            "items": {
              "$ref": "#/components/schemas/ProjectView"
            },
            "title": "Projects",
            "type": "array"
          }
        },
        "required": [
          "projects"
        ],
        "title": "ProjectList",
        "type": "object"
      },
      "ProjectView": {
        "additionalProperties": false,
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "created_at"
        ],
        "title": "ProjectView",
        "type": "object"
      },
...
      "Spec2WavManifest": {
        "additionalProperties": false,
        "properties": {
          "complete": {
            "const": true,
            "default": true,
            "title": "Complete",
            "type": "boolean"
          },
          "core_version": {
            "title": "Core Version",
            "type": "string"
          },
          "files": {
            "items": {
              "$ref": "#/components/schemas/AcousticManagedFile"
            },
            "maxItems": 4,
            "minItems": 4,
            "title": "Files",
            "type": "array"
          },
          "kind": {
            "const": "managed_spec2wav_files",
            "default": "managed_spec2wav_files",
            "title": "Kind",
            "type": "string"
          },
          "operation": {
            "const": "spectrogram_to_audio",
            "default": "spectrogram_to_audio",
            "title": "Operation",
            "type": "string"
          },
          "policy_version": {
            "default": 1,
            "enum": [
              1,
              2
            ],
            "title": "Policy Version",
            "type": "integer"
          }
        },
        "required": [
          "core_version",
          "files"
        ],
        "title": "Spec2WavManifest",
        "type": "object"
      },
...
      "UploadInput": {
        "additionalProperties": false,
        "properties": {
          "expected_bytes": {
            "anyOf": [
              {
                "maximum": 1000000000.0,
                "minimum": 0.0,
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Bytes"
          },
          "idempotency_key": {
            "maxLength": 128,
            "minLength": 16,
            "pattern": "^[A-Za-z0-9_-]+$",
            "title": "Idempotency Key",
            "type": "string"
          },
          "name": {
            "maxLength": 180,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          },
          "project_id": {
            "format": "uuid",
            "title": "Project Id",
            "type": "string"
          }
        },
        "required": [
          "project_id",
          "name",
          "idempotency_key"
        ],
        "title": "UploadInput",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Shared API; accounts require configured PostgreSQL. M01 batches require an explicitly configured durable worker and resource store.",
    "title": "PhoneticToolbox API",
    "version": "1.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/assets": {
      "get": {
        "operationId": "list_assets",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "expires",
              "enum": [
                "expires",
                "size",
                "created"
              ],
              "title": "Order",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listing",
        "tags": [
          "storage"
        ]
      }
    },
...
    "/api/v1/jobs": {
      "get": {
        "operationId": "list_jobs",
        "parameters": [
          {
            "in": "query",
            "name": "project_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Project Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Listing",
        "tags": [
          "jobs"
        ]
      },
      "post": {
        "operationId": "create_job",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "anyOf": [
                  {
                    "$ref": "#/components/schemas/JobInput"
                  },
                  {
                    "$ref": "#/components/schemas/FileJobInput"
                  }
                ],
                "title": "Body"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "jobs"
        ]
      }
    },
...
    "/api/v1/projects": {
      "get": {
        "operationId": "list_projects",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "Listing",
        "tags": [
          "projects"
        ]
      },
      "post": {
        "operationId": "create_project",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "projects"
        ]
      }
    },
...
    "/api/v1/uploads": {
      "post": {
        "operationId": "create_upload",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UploadInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssetView"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create",
        "tags": [
          "storage"
        ]
      }
    },
...
  }
}